design landing page for the mobile
