+++ src/components/TrialPageComponents/CanvasDemo.tsx
@@ -2,12 +2,13 @@ import React from "react";
 import Image from "next/image";
 
 import canvasGif from "../../../public/canvas.gif";
+import canvasMobileGif from "../../../public/canvas-mobile.gif";
 import DemoContainer from "./DemoContainer";
 
 function CanvasDemo() {
 	return (
-		<DemoContainer>
-			<div className="flex w-fit border-rose-400 border-2 rounded-md mr-20 drop-shadow-red">
+		<DemoContainer reverse={true}>
+			<div className="hidden lg:flex w-fit border-rose-400 border-2 rounded-md drop-shadow-red">
 				<Image
 					src={canvasGif.src}
 					width={604}
@@ -15,7 +16,15 @@ function CanvasDemo() {
 					layout="fixed"
 				/>
 			</div>
-			<div className="flex gap-2 items-center text-3xl tracking-wide font-bold text-white  h-fit">
+			<div className="flex flex-col lg:hidden  border-rose-400 border-2 rounded-md  drop-shadow-red">
+				<Image
+					src={canvasMobileGif.src}
+					width={320}
+					height={605}
+					layout="fixed"
+				/>
+			</div>
+			<div className="flex gap-2 items-center text-2xl lg:text-3xl tracking-wide font-bold text-white  h-fit">
 				<span>
 					Draw{" "}
 					<span className="text-rose-400">Free Hand Diagrams! </span>
