+++ src/pages/index.tsx
@@ -1,139 +1,85 @@
-import type { GetStaticProps, NextPage } from "next";
-import Head from "next/head";
+import { GetStaticProps } from "next";
+import Link from "next/link";
 import { useRouter } from "next/router";
-import { useContext, useEffect, useState } from "react";
-import {
-	LIMIT,
-	SEARCH_PUBLC,
-	SUPABASE_POST_TABLE,
-} from "../../utils/constants";
-import { fetchUpvotes } from "../../utils/fetchUpvotes";
+import { useContext, useEffect } from "react";
+import { AiFillRead } from "react-icons/ai";
+import { FaPencilAlt } from "react-icons/fa";
+import { SUPABASE_FILES_BUCKET } from "../../utils/constants";
 import { supabase } from "../../utils/supabaseClient";
 import Layout from "../components/Layout";
-import PostDisplay from "../components/PostDisplay";
-import SearchComponent from "../components/SearchComponent";
-import PostWithBlogger from "../interfaces/PostWithBlogger";
-import SearchResults from "../interfaces/SearchResult";
+import AllScreenDemo from "../components/TrialPageComponents/AllScreenDemo";
+import CanvasDemo from "../components/TrialPageComponents/CanvasDemo";
+import HeaderText from "../components/TrialPageComponents/HeaderText";
+import MdToBlog from "../components/TrialPageComponents/MdToBlogDemo";
+import SearchDemo from "../components/TrialPageComponents/SearchDemo";
+import WriteEditDemo from "../components/TrialPageComponents/WriteEditDemo";
 import { UserContext } from "./_app";
+// import markdownArray from "../../utils/trialArray";
 
-interface HomeProps {
-	posts: PostWithBlogger[] | null;
+interface TrialProps {
+	markdown: string;
 }
-const Home: NextPage<HomeProps> = ({ posts }) => {
-	const router = useRouter();
+
+function Trial({ markdown }: TrialProps) {
 	const { user } = useContext(UserContext);
-	const [homePosts, setHomePosts] = useState<
-		Partial<PostWithBlogger>[] | null | undefined
-	>(posts);
-	const [searchResults, setSearchResults] = useState<SearchResults[]>();
-	const [searchQuery, setSearchQuery] = useState("");
+	const router = useRouter();
 
 	useEffect(() => {
-		fetchUpvotes(homePosts, setHomePosts);
+		const observer = new IntersectionObserver(
+			(entries) => {
+				entries.forEach((entry) => {
+					if (entry.isIntersecting) {
+						entry.target.classList.add("opaque");
+						return;
+					}
+					// entry.target.classList.remove('opaque')
+				});
+			},
+			{ threshold: 0.5 }
+		);
+		const transparentElements = document.querySelectorAll(".transparent");
+		transparentElements.forEach((el) => observer.observe(el));
 	}, []);
 
-	const fetchHomePosts = async ({ cursor }: { cursor: string | number }) => {
-		const { data, error } = await supabase
-			.from<PostWithBlogger>(SUPABASE_POST_TABLE)
-			.select("*,bloggers(name)")
-			.match({ published: true })
-			.lt("published_on", cursor)
-			.order("published_on", { ascending: false })
-			.limit(LIMIT);
-
-		if (error || !data) {
-			console.log(error.message || "data returned is null");
-			return false;
-		}
-		setHomePosts((prev) => [...(prev || []), ...data]);
-		return data.length > 0;
-	};
-
-	const fetchSearchPosts = async ({
-		cursor,
-		searchTerm,
-	}: {
-		cursor?: string | number;
-		searchTerm?: string;
-	}) => {
-		if (!searchTerm) return;
-
-		const { data, error } = await supabase.rpc<SearchResults>(
-			SEARCH_PUBLC,
-			{
-				search_term: searchTerm,
-				cursor: cursor || null,
-			}
-		);
-
-		if (error || !data) return false;
-
-		setSearchResults(data);
-		return data.length > 0;
-	};
 	return (
 		<Layout user={user || null} route={router.asPath}>
-			<Head>
-				<title>Rce Blog</title>
-				<meta
-					name="description"
-					content="Write notes containing prose, executable code snippets, free hand drawings and images."
-				/>
-				<meta
-					name="keywords"
-					content="RCE-Blog, remote code execution, blog, tech blog"
-				/>
-				<meta
-					property="og:title"
-					content="Home Page of the RCE-Blog website"
-				/>
-				<meta property="og:url" content="https://rce-blog.xyz" />
-				<meta property="og:site_name" content="RCE-Blog" />
-				<meta property="og:type" content="website" />
-			</Head>
-			<div className="w-full px-5 lg:px-0 md:w-3/5 lg:w-1/3 mx-auto">
-				<SearchComponent
-					fetchPosts={fetchSearchPosts}
-					setPosts={setSearchResults}
-					setSearchQuery={setSearchQuery}
-				/>
-			</div>
-			<div className="px-4 lg:px-32 xl:px-64 grow mt-12 overflow-hidden">
-				{(searchResults?.length || 0) > 0 ? (
-					<PostDisplay
-						posts={searchResults || []}
-						cursorKey="search_rank"
-						searchTerm={searchQuery}
-						fetchPosts={fetchSearchPosts}
-					/>
-				) : (
-					<PostDisplay
-						posts={homePosts || []}
-						cursorKey="published_on"
-						searchTerm={searchQuery}
-						fetchPosts={fetchHomePosts}
-					/>
-				)}
+			<div className="flex gap-10 lg:gap-20 md:gap-14 flex-col pb-20 grow overflow-y-auto">
+				<div className="self-center flex gap-10 font-semibold lg:text-xl text-black">
+					<Link href="/read">
+						<div className="cursor-pointer bg-amber-400 flex items-center gap-2 px-3 lg:px-4 py-1 lg:py-2 rounded-md  ">
+							Read <AiFillRead size={24} />
+						</div>
+					</Link>
+					<Link href="/edit">
+						<div className="cursor-pointer px-3 lg:px-4 py-1 lg:py-2 rounded-md   flex items-center gap-2 bg-amber-400">
+							Write <FaPencilAlt />
+						</div>
+					</Link>
+				</div>
+				<HeaderText />
+				<MdToBlog {...{ markdown }} />
+				<WriteEditDemo />
+				<CanvasDemo />
+				<SearchDemo />
+				<AllScreenDemo />
 			</div>
 		</Layout>
 	);
-};
+}
 
-export const getStaticProps: GetStaticProps<HomeProps> = async ({}) => {
-	const { data } = await supabase
-		.from<PostWithBlogger>(SUPABASE_POST_TABLE)
-		.select(
-			`id,created_by,title,description,language,published,published_on,bloggers(name)`
-		)
-		.match({ published: true })
-		.order("published_on", { ascending: false })
-		.limit(LIMIT);
+export const getStaticProps: GetStaticProps<TrialProps> = async ({}) => {
+	const { data: fileData, error: fileError } = await supabase.storage
+		.from(SUPABASE_FILES_BUCKET)
+		.download("f2c61fc8-bcdb-46e9-aad2-99c0608cf485/608/file.md");
 
+	if (fileError || !fileData)
+		return { props: { markdown: "" }, redirect: "/" };
+	const content = await fileData.text();
 	return {
 		props: {
-			posts: data,
+			markdown: content,
 		},
 	};
 };
 
-export default Home;
+export default Trial;
