+++ utils/todo.txt
@@ -3,6 +3,8 @@
 4. maybe canvas should be excalidraw
 5. Figure out activate rce ux for phone
 6. Comma should not be the caption separator in carousels.
+7. Design landing page for small screens
+
 
 -----------------Done------------------------------
 2. Modify carousel UX.
